xwayland: Handle the wl_keyboard modifiers event
This allows us to keep track of latched and locked modifiers as well
as the XKB group while the keyboard focus is on another wayland
client. Note that we don't need to track depressed modifiers since
wayland guarantees that we'll get them as key press events whenever
we get the keyboard focus.